+++ tests/isolation_provider/test_qubes.py
@@ -1,0 +1,15 @@
+import pytest
+
+from dangerzone.isolation_provider.qubes import Qubes
+
+from .. import sanitized_text, uncommon_text
+from .base import IsolationProviderTest
+
+
+@pytest.fixture
+def provider() -> Qubes:
+    return Qubes()
+
+
+class TestQubes(IsolationProviderTest):
+    pass
